Corect VAO List, With generation and deletion
Fix the list of vaos. Vaos can now be deleted from the list.
Implement Shader functions. Two types will be supported.
Vertex- and fragmentshaders. Both take in as parameters the
number of attributes, the actual attributes passed as an
array of union vectors (mixed sizes of 1, 2, 3 or 4) and the
4-component vector output, which is either the vertex position,
or the computed color.