Attribute based rendering with vertex shader
Implement attribute based rendering. For each index the
corresponding attributes are collected and passed to a vertex
shader callback of the following signature:

SR_Vec4f vertex_shader (size_t count, SR_VecUnion *attribs).

The parameters are the number of attributes and the attribute array.
The attribute indices correspond to the indices set by
SR_SetVertexAttribute. The attributes are pulled from the vertexdata,
according to the parameters set for the index in SR_SetVertexAttribute.

The return value should be the final vertex position in NDC space.