Vertex Stage Output, missing interpolation
It is now possible to feed additional vertex attributes and form
the vertex stage output by first declaring the wanted output size
(so the software renderer can allocate storage) and the filling the
vertex stage output (at the index positions) with the desired values
in the vertex shader (or anywhere else).

A pointer to the collected per patch attributes is passed further down
the pipeline.